- Fixed build issue in autogenerated code where (double) was removed when it shouldn't have been

+++ main/ejml-zdense/src/org/ejml/dense/row/NormOps_ZDRM.java
@@ -54,6 +54,6 @@ public class NormOps_ZDRM {
             total += real*real + imag*imag;
         }
 
-        return scale*Math.sqrt(total);
+        return scale * (double)Math.sqrt(total);
     }
 }
